python-service/requirements.txt
# Core
numpy==1.25.2
faiss-cpu==1.12.0
sentence-transformers==2.7.0
requests==2.31.0
python-dotenv==1.0.0
pymongo==4.6.0

# FastAPI for API service
fastapi==0.116.1
uvicorn==0.35.0

# Data dependencies
pandas==2.3.2
scikit-learn==1.7.1
pyarrow==21.0.0

# HuggingFace dependencies
torch==2.3.0
torchvision==0.17.1
transformers==4.44.2
